+++ src/main/java/de/ellpeck/actuallyadditions/mod/blocks/BlockCrusher.java
@@ -100,18 +100,4 @@ public class BlockCrusher extends BlockContainerBase {
                 ? 12
                 : 0;
     }
-
-/*    @Override
-    public VoxelShape getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
-        switch (state.getValue(HORIZONTAL_FACING)) {
-            case EAST:
-                return VoxelShapes.GrinderShapes.SHAPE_E;
-            case SOUTH:
-                return VoxelShapes.GrinderShapes.SHAPE_S;
-            case WEST:
-                return VoxelShapes.GrinderShapes.SHAPE_W;
-            default:
-                return VoxelShapes.GrinderShapes.SHAPE_N;
-        }
-    }*/
 }