+++ src/main/java/de/ellpeck/actuallyadditions/mod/blocks/BlockDropper.java
@@ -50,22 +50,4 @@ public class BlockDropper extends FullyDirectionalBlock.Container {
 
         return this.openGui(world, player, pos, TileEntityDropper.class);
     }
-
-/*    @Override
-    public VoxelShape getShape(BlockState state, BlockGetter worldIn, BlockPos pos, CollisionContext context) {
-        switch (state.getValue(FACING)) {
-            case UP:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_U;
-            case DOWN:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_D;
-            case EAST:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_E;
-            case SOUTH:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_S;
-            case WEST:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_W;
-            default:
-                return VoxelShapes.BlockBreakerShapes.SHAPE_N;
-        }
-    }*/
 }